[PATCH] sign fix in swapfile.c
  CC      mm/swapfile.o
mm/swapfile.c: In function `scan_swap_map':
mm/swapfile.c:114: warning: comparison between signed and unsigned

+++ mm/swapfile.c
@@ -110,7 +110,7 @@ static inline int scan_swap_map(struct swap_info_struct *si)
  check_next_cluster:
 	if (offset+SWAPFILE_CLUSTER-1 <= si->highest_bit)
 	{
-		int nr;
+		unsigned long nr;
 		for (nr = offset; nr < offset+SWAPFILE_CLUSTER; nr++)
 			if (si->swap_map[nr])
 			{
